reinstall extension package

+++ app/code/Magefan/TranslationPlus/Controller/Adminhtml/Category/Save.php
@@ -4,7 +4,6 @@
  * Please visit Magefan.com for license details (https://magefan.com/end-user-license-agreement).
  */
 
-
 namespace Magefan\TranslationPlus\Controller\Adminhtml\Category;
 
 use Magento\Framework\Exception\LocalizedException;
@@ -16,6 +15,7 @@ use Magefan\Translation\Api\ConfigInterface;
 use Magento\Framework\Message\ManagerInterface;
 use Magento\Catalog\Model\ResourceModel\Category\CollectionFactory;
 use Magento\Framework\Controller\Result\JsonFactory;
+use Magento\Framework\Event\ManagerInterface as EventManager;
 
 class Save extends \Magento\Backend\App\Action
 {
@@ -54,14 +54,20 @@ class Save extends \Magento\Backend\App\Action
      */
     private $resultJsonFactory;
 
+    private $eventManager;
+
     /**
-     * Save constructor.
+     * @var array
+     */
+    private $use301redirect = [];
+
+    /**
      * @param Context $context
      * @param StoreRepositoryInterface $repository
      * @param CategoryRepository $categoryRepository
      * @param ConfigInterface $configInterface
      * @param ManagerInterface $messageManager
-     * @param CategoryFactory $categoryCollectionFactory
+     * @param CollectionFactory $categoryCollectionFactory
      * @param JsonFactory $resultJsonFactory
      */
     public function __construct(
@@ -71,7 +77,8 @@ class Save extends \Magento\Backend\App\Action
         ConfigInterface $configInterface,
         ManagerInterface $messageManager,
         CollectionFactory $categoryCollectionFactory,
-        JsonFactory $resultJsonFactory
+        JsonFactory $resultJsonFactory,
+        EventManager $eventManager
     ) {
         $this->repository = $repository;
         $this->categoryRepository = $categoryRepository;
@@ -79,7 +86,8 @@ class Save extends \Magento\Backend\App\Action
         $this->messageManager = $messageManager;
         $this->categoryCollectionFactory = $categoryCollectionFactory;
         $this->resultJsonFactory = $resultJsonFactory;
-        return parent::__construct($context);
+        $this->eventManager = $eventManager;
+        parent::__construct($context);
     }
 
     /**
@@ -120,6 +128,8 @@ class Save extends \Magento\Backend\App\Action
         $post[0] = $post[-1];
         unset($post[-1]);
 
+        $this->collect301Redirects($post);
+
         try {
             $model = $this->categoryCollectionFactory->create();
             $resource     = $model->getResource();
@@ -129,12 +139,25 @@ class Save extends \Magento\Backend\App\Action
                 if ($store->getId()) {
                     foreach ($post[0] as $attrKey => $value) {
                         if (!isset($post[$store->getId()][$attrKey])) {
+                            if ($attrKey == 'url_key') {
+                                // in case when checkbox Use Default Checked
+                                $post[$store->getId()]['url_key'] = false;
+                                $post[$store->getId()]['url_key_create_redirect'] = $this->use301redirect[$store->getId()] ?? 0;
+                                continue;
+                            }
+
                             $attribute = $resource->getAttribute($attrKey);
+
+                            if (!$attribute) {
+                                continue;
+                            }
+
                             $condition = [
                                 'attribute_id = ?'   => $attribute->getId(),
                                 'store_id = ?'       => $store->getId(),
                                 'entity_id IN(?)'    => $id
                             ];
+
                             $adapter->delete($attribute->getBackendTable(), $condition);
                         }
                     }
@@ -142,6 +165,7 @@ class Save extends \Magento\Backend\App\Action
             }
 
             ksort($post);
+
             foreach ($post as $storeId => $value) {
                 if (!is_numeric($storeId)) {
                     continue;
@@ -150,23 +174,93 @@ class Save extends \Magento\Backend\App\Action
                 $category = $this->categoryRepository->get($id, (int)$storeId);
 
                 foreach ($post[$storeId] as $attrKey => $attrValue) {
-
-                    if ($attrKey == 'category_has_weight') {
-                        continue;
-                    }
                     if (is_array($attrValue)) {
                         $attrValue = implode(',', $attrValue);
                     }
 
-                    $category->setData($attrKey, $attrValue);
-                    $category->getResource()->saveAttribute($category, $attrKey);
+                    switch ($attrKey) {
+                        case "url_key_create_redirect":
+                        case "category_has_weight":
+                        case "locale_code":
+                        case "auto_translation_original":
+                            break;
+                        case "url_key":
+                            $this->saveUrlKey($category, $post[$storeId], $storeId, $model);
+                            break;
+                        default:
+                            $category->setData($attrKey, $attrValue);
+                            $category->getResource()->saveAttribute($category, $attrKey);
+                    }
                 }
             }
+            $this->eventManager->dispatch('magefan_save_catalog_translation', ['current_object' => $this]);
             return $this->processResponse(true, __('You saved the category.'));
         } catch (LocalizedException $e) {
             return $this->processResponse(false, $e->getPrevious() ? : $e);
         } catch (\Exception $err) {
             return $this->processResponse(false, __('Something went wrong while saving the category.'));
+        }
+    }
+
+    /**
+     * @param $category
+     * @param $data
+     * @param $storeId
+     * @param $model
+     * @return void
+     * @throws NoSuchEntityException
+     */
+    private function saveUrlKey($category, $data, $storeId, $model)
+    {
+        $oldUrl = $category->getUrlKey();
+        $newUrl = $data['url_key'];
+
+        if (($newUrl || $newUrl === '') && $oldUrl != $newUrl) {
+
+            // in case of empty string - url key will be generated based on entity name
+            if ($newUrl === '') {
+                $newUrl = null;
+            }
+
+            $category->setData('url_key_create_redirect', '');
+            $category->setData('save_rewrites_history', false);
+
+            if (isset($data['url_key_create_redirect']) && $data['url_key_create_redirect']) {
+                $category->setData('url_key_create_redirect', $oldUrl);
+                $category->setData('save_rewrites_history', true);
+            }
+
+            $category->setUrlKey($newUrl);
+            $this->_eventManager->dispatch('catalog_category_save_before', ['category' => $category]);
+            $category->getResource()->saveAttribute($category, 'url_key');
+
+            $this->_eventManager->dispatch('catalog_category_save_after', ['category' => $category]);
+        } elseif ($newUrl === false && $category->getExistsStoreValueFlag('url_key')) {
+            // in case when use_default checked
+
+            $category->setData('url_key_create_redirect', '');
+            $category->setData('save_rewrites_history', true);
+
+            $categoryBaseStore = $this->categoryRepository->get($category->getData('entity_id'), 0);
+
+            $category->setUrlKey($categoryBaseStore->getUrlKey());
+
+            $this->_eventManager->dispatch('catalog_category_save_before', ['category' => $category]);
+            $category->getResource()->saveAttribute($category, 'url_key');
+            $this->_eventManager->dispatch('catalog_category_save_after', ['category' => $category]);
+
+            // remove value from store level
+            $resource = $model->getResource();
+            $adapter = $resource->getConnection();
+            $attribute = $resource->getAttribute('url_key');
+
+            $condition = [
+                'attribute_id = ?'   => $attribute->getId(),
+                'store_id = ?'       => $storeId,
+                'entity_id IN(?)'    => $category->getData('entity_id')
+            ];
+
+            $adapter->delete($attribute->getBackendTable(), $condition);
         }
     }
 
@@ -196,7 +290,7 @@ class Save extends \Magento\Backend\App\Action
                     ? $this->getUrl('catalog/category/edit', ['id' => $id])
                     : $this->getUrl('translationplus/category/translate', ['id' => $id]);
             } else {
-                $this->messageManager->addExceptionMessage($message);
+                $this->messageManager->addErrorMessage($message);
 
                 $path = ($id)
                     ? $this->getUrl('translationplus/category/translate', ['id' => $id])
@@ -205,6 +299,25 @@ class Save extends \Magento\Backend\App\Action
             }
 
             return $resultRedirect->setPath($path);
+        }
+    }
+
+    /**
+     * @param array $post
+     */
+    private function collect301Redirects(array &$post): void
+    {
+        $useRedirectForDefaultValue = $post['use_redirect_for_default_value'] ?? [];
+
+        if ($useRedirectForDefaultValue) {
+
+            foreach ($useRedirectForDefaultValue as $storeId_isChecked) {
+                [$storeId, $isChecked] = explode('_', $storeId_isChecked);
+
+                $this->use301redirect[$storeId] = $isChecked;
+            }
+
+            unset($post['use_redirect_for_default_value']);
         }
     }
 }
